Test de logging + ajout d'un module et d'un module pour scrap les matchs
