+++ .gitignore
@@ -11,3 +11,6 @@ wheels/
 idee.md
 
 ml-webscrap-tennis.bat
+
+log-example.log
+logging-example.py
