+++ logs/scraping_donnees_joueurs.log
@@ -1,0 +1,24 @@
+2024-12-29 14:50:20,431 - INFO - Script de scraping des données des joueurs démarré.
+2024-12-29 14:50:20,440 - INFO - Fichier joueurs chargé avec succès.
+2024-12-29 14:50:20,449 - INFO - Fichier D:\Cours_M2\webscraping\ml-webscrap-tennis\data\detail_joueurs.json chargé avec succès.
+2024-12-29 14:50:20,449 - INFO - Début du scraping pour Jannik_Sinner...
+2024-12-29 14:50:20,897 - INFO - Données de Jannik_Sinner sauvegardées avec succès.
+2024-12-29 14:50:22,898 - INFO - Début du scraping pour Alexander_Zverev...
+2024-12-29 14:50:23,292 - INFO - Données de Alexander_Zverev sauvegardées avec succès.
+2024-12-29 14:50:25,301 - INFO - Début du scraping pour Carlos_Alcaraz...
+2024-12-29 14:50:25,716 - INFO - Données de Carlos_Alcaraz sauvegardées avec succès.
+2024-12-29 14:50:27,717 - INFO - Début du scraping pour Taylor_Harry_Fritz...
+2024-12-29 14:50:28,179 - INFO - Données de Taylor_Harry_Fritz sauvegardées avec succès.
+2024-12-29 14:50:30,179 - INFO - Début du scraping pour Daniil_Medvedev...
+2024-12-29 14:50:30,614 - INFO - Données de Daniil_Medvedev sauvegardées avec succès.
+2024-12-29 14:50:32,627 - INFO - Début du scraping pour Casper_Ruud...
+2024-12-29 14:50:33,022 - INFO - Données de Casper_Ruud sauvegardées avec succès.
+2024-12-29 14:50:35,023 - INFO - Début du scraping pour Novak_Djokovic...
+2024-12-29 14:50:35,609 - INFO - Données de Novak_Djokovic sauvegardées avec succès.
+2024-12-29 14:50:37,610 - INFO - Début du scraping pour Andrey_Rublev...
+2024-12-29 14:50:37,981 - INFO - Données de Andrey_Rublev sauvegardées avec succès.
+2024-12-29 14:50:39,982 - INFO - Début du scraping pour Alex_De_Minaur...
+2024-12-29 14:50:40,525 - INFO - Données de Alex_De_Minaur sauvegardées avec succès.
+2024-12-29 14:50:42,525 - INFO - Début du scraping pour Grigor_Dimitrov...
+2024-12-29 14:50:43,027 - INFO - Données de Grigor_Dimitrov sauvegardées avec succès.
+2024-12-29 14:50:45,028 - INFO - Tous les joueurs ont été traités et sauvegardés dans D:\Cours_M2\webscraping\ml-webscrap-tennis\data\detail_joueurs.json.
