+++ run_path.bat
@@ -1,0 +1,2 @@
+@echo off
+set PYTHONPATH=D:\Cours_M2\webscraping\ml-webscrap-tennis
